async-aggregators/pyproject.toml
[tool.poetry]
name = "Name"
version = "0.1.0"
description = ""
authors = ["Author 1 <email@mail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "3.10.*"
scikit-learn = "^1.3.0"
pyqt6 = "~6.5.2"
pillow = "~10.0.0"
numpy = "~1.25.2"
opencv-python-headless = "^4.8.0.76"


[tool.poetry.dev-dependencies]
# Base tools
flake8 = "~4.0.1"
isort = "~5.12.0"
pre-commit = "~3.3.3"
mypy = "~1.5.1"
flake8-docstrings = "~1.6.0"
pillow = "~10.0.0"
numpy = "~1.25.2"
scikit-learn = "~1.3.0"
pyqt6 = "~6.5.2"



[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"